Initial specs
This is an initial batch of specs written using RSpec. Seeing as this is
such a simple tool there doesn't seem to be much sense in doing internal
testing at the Objective-C level; instead just run the tool itself and
make sure it does what it's supposed to do. RSpec is used despite the
fact that wincent-strings-util has absolutely no Ruby in it, simply
because it's a nice framework for doing this kind of thing.

Signed-off-by: Wincent Colaiuta <win@wincent.com>