distro.sh: distribute package file without zip wrapper
I've now configured the web servers (both nginx and Apache) to serve
".pkg" files with a MIME type of "application/octet-stream" instead of
the default "text/plain". This means that it is no longer necessary to
wrap them up inside zip archives to prevent them from being
inappropriately displayed in the user's browser window rather than
downloaded to disk.

So, drop the zip wrapper and instead just give the package a unique
name. Flat packages are compressed in any case, so the download size
should be basically the same, give or take a few bytes.

This commit is the equivalent of the change made in 17fb9e0 in Wincent
Icon Utility.

Signed-off-by: Wincent Colaiuta <win@wincent.com>